TITANIC SURVIVAL PREDICTION
The Titanic prediction model aims to predict the survival of passengers on the Titanic using their demographic and ticket information. By leveraging the Titanic dataset, we preprocess and clean the data, converting categorical variables into numerical ones. We then split the data into training and testing sets. Using logistic regression, a binary classification algorithm, we train a model to distinguish between survivors and non-survivors. Finally, we evaluate the model's performance with metrics like accuracy and the confusion matrix to assess its predictive capabilities.